Added static variable PATTERN with regular expression "\\p{IsPunctuation}" that covers punctuation marks.
The readFromFile() method takes the string fileName and internally implements the File object that is passed to the readFiles() method.

The readFiles() method has been added, in which a StringBuilder is declared in which the result of reading the file will be written, BufferedReader, which reads the file and String value to store the string before line wrapping, since this method does not read the entire text at once, but line by line.
After that, a while loop is declared, inside which in StringBuilder append the contents of value, after which value is updated with the next line.
Next, the method sends the finished StringBuilder to the wordsWFilter() method.
The entire content of the readFiles() method runs in a try-catch block in case we get an IOException if we can't read the file.

The wordsWFilter() method has been added, in which the length of the StringBuilder is checked first. In case of 0, we return an empty String[] array.
Otherwise, we create a String[] array, in which:
- replace all punctuation marks with a space
- remove extra spaces
- Convert all characters to lowercase
- Separate words by spaces.
Then we reset StringBuilder. We'll still need it.
After that, we loop through the resulting array, and if the first character of the array element is 'w', then we add this entire string to the StringBuilder and separate the elements with the ":" character.
Next, we pass the result as a String[] array to the arraySort() method, separating the elements by the ":" symbol, but if the StringBuilder is empty after the loop, then we return an empty String[] array.

Added the arraySort() method, which sorts the contents of the String[] array. For this, two for loops are used. The inner (second) loop avoids repetition when comparing. If the condition (array[i].compareTo(array[j])>0) is true then 0, it performs the swapping and sorts the array.
The resulting result is returned.